more passing tests

+++ uk.ac.diamond.ispyb.api/src/uk/ac/diamond/ispyb/api/IspybDataCollectionApi.java
@@ -3,8 +3,8 @@ package uk.ac.diamond.ispyb.api;
 import java.io.Closeable;
 
 public interface IspybDataCollectionApi extends Closeable {
-	public long upsertDataCollectionMain(DataCollectionMain dataCollectionMain);
+	public int upsertDataCollectionMain(DataCollectionMain dataCollectionMain);
 	public void updateDataCollectionExperiment(DataCollectionExperiment dataCollectionExperiment);
 	public void updateDataCollectionMachine(DataCollectionMachine dataCollectionMachine);
-	public long upsertDataCollectionGroup(DataCollectionGroup dataCollectionGroup);
+	public int upsertDataCollectionGroup(DataCollectionGroup dataCollectionGroup);
 }
